Typescript compatibility example.
