+++ examples/typescript-compatibility/README.md
@@ -1,0 +1,13 @@
+# Typescript Compatibility Example
+
+This example demonstrates how to use GenSX packages with an older configuration of Typescript that uses `moduleResolution: "Node"` and `target: "es5"`.
+
+## The Example
+
+This example has typescript configured with `moduleResolution: "Node"` and `target: "es5"`.
+
+## Running the typescript example
+
+```bash
+pnpm start
+```
