+++ examples/typescript-compatibility/eslint.config.mjs
