+++ examples/typescript-compatibility/nodemon.json
