+++ examples/typescript-compatibility/package.json
@@ -7,21 +7,21 @@
   },
   "scripts": {
     "dev": "nodemon",
-    "start:ts": "tsx index.tsx",
-    "start": "node ./dist/index.js",
-    "start:pure": "node ./pure-commonjs.js",
+    "start": "tsx index.tsx",
     "build": "tsc",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix"
   },
   "dependencies": {
     "@gensx/core": "workspace:*",
+    "@gensx/openai": "workspace:*",
+    "openai": "^4.88.0",
     "zod": "catalog:"
   },
   "devDependencies": {
     "@types/node": "catalog:examples",
     "nodemon": "catalog:",
     "tsx": "catalog:",
-    "typescript": "catalog:"
+    "typescript": "^5.0.0"
   }
 }
